Move relation actions and EnvVarViewComponent into cf module
