+++ src/frontend/packages/cloud-foundry/src/shared/components/env-var-view/env-var-view.component.scss
