+++ src/frontend/packages/core/src/shared/components/application-state/application-state.component.ts
@@ -4,7 +4,6 @@ import { map, startWith } from 'rxjs/operators';
 
 import { StratosStatus, StratosStatusMetadata } from '../../shared.types';
 
-// TODO: RC MOVE
 @Component({
   selector: 'app-application-state',
   templateUrl: './application-state.component.html',
